+++ my-scheme-compiler/src/lib.rs
@@ -1,4 +1,4 @@
 pub mod compile;
-pub mod compilers;
-pub mod core;
 pub mod error;
+pub mod module_compiler;
+pub mod symbol_list;
