+++ my-scheme-compiler/tests/test_program/Cargo.toml
@@ -4,4 +4,4 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-my-scheme-lib = { path="../../" }
+my-scheme-lib = { path="../../../my-scheme-lib" }
