+++ my-scheme-compiler/tests/test_program/src/main.rs
@@ -1,20 +1,15 @@
-use my_scheme_lib::compile::core::*;
-use my_scheme_lib::data::err::Error;
 use my_scheme_lib::data::value::Value;
+use my_scheme_lib::data::err::Error;
+use my_scheme_lib::proc::help::*;
 use my_scheme_lib::list;
-fn program() -> Result<Value, Error> {
-    let __1 = symbol("a");
-    let __2 = symbol("+");
-    let env = new_env();
-    Ok(Value::from(vec![
-        Value::from(1),
-        list![Value::from(4), __1.clone(), __2.clone()],
-        Value::from(3),
-    ]))
+pub fn module(env: Environment) -> Result<Value, Error> {
+{ let env = push(&env); put(&env, "f", Value::Undefined); put(&env, "g", Value::Undefined); set(&env, "f", lambda(None, env.clone(), |args, env| { put(&env, "x", args[0].clone()); apply( get(&env, "g")?, vec![get(&env, "x")?] ) })?)?; set(&env, "g", lambda(None, env.clone(), |args, env| { put(&env, "x", args[0].clone()); apply( get(&env, "+")?, vec![Value::from(1), get(&env, "x")?] ) })?)?; apply( get(&env, "f")?, vec![Value::from(6)] ) }
 }
+
 fn main() {
-    match program() {
+    let env = new_env();
+    match module(env) {
         Ok(val) => println!("{val}"),
         Err(e) => println!("{:?}", e),
-    }
+   }
 }
